# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: 通学AIラジオ API
  version: "0.1.0"
  description: |
    フロントエンド（React + Vite）とバックエンド（Go）を並行開発するための取り決め。
    **実装より先にこのファイルを更新して合意すること。** 変更したら両チームに共有する。

    - データモデルは [06-storage.md](../backend/docs/pipeline/06-storage.md) の PostgreSQL スキーマに対応する
    - 認証は行わない（ログイン機能はスコープ外。`userId` の発行方法は下記「userId の扱い」参照）。
      ローカル完結・オフライン前提の構成であり、本番公開は想定していないための割り切り

    ## CORS
    フロントエンド（`http://localhost:3000` 想定）とバックエンド（`http://localhost:8080` 想定）は
    オリジンが異なるため、バックエンドは以下の CORS レスポンスヘッダを返すこと。
    実装時に抜けやすく、抜けると疎通確認の最初の一歩で詰まるので明記しておく。

    ```
    Access-Control-Allow-Origin: http://localhost:3000
    Access-Control-Allow-Methods: GET, POST, PUT, OPTIONS
    Access-Control-Allow-Headers: Content-Type
    ```

    ## userId の扱い
    ログイン機能はスコープ外のため、次の方式で暫定的にユーザーを識別する。

    1. フロントエンドは初回起動時に `POST /api/users` を呼び、サーバーが発行した `userId`（UUID）を受け取る
    2. `userId` をブラウザの `localStorage` に保存し、以降の全リクエストで使い回す
    3. サーバーはリクエストされた `userId` が `users` テーブルに存在するかを検証し、
       存在しなければ `404`（`USER_NOT_FOUND`）を返す

    クライアントが自前で UUID を生成しない理由：`programs.user_id` が外部キー制約付きのため、
    存在しないIDでの書き込みを防ぐにはサーバー発行が確実。

    ## 表示名（ニックネーム）の扱い
    UIモック（`/mock`）ではプロフィール画面にユーザー名を表示するが、ログイン機能がスコープ外である以上
    サーバー側でアカウントに紐づく表示名を管理する意味は薄い。表示名はバックエンドAPIの対象外とし、
    フロントエンドの `localStorage` にのみ保存するクライアントローカルな値として扱う
    （どのAPIレスポンスにも `name` 相当のフィールドは含まれない）。
servers:
  - url: http://localhost:8080
    description: ローカル開発環境（VITE_API_BASE_URL、詳細は frontend/CLAUDE.md#環境変数）
...
paths:
...
  /api/users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: ユーザープロフィールを取得する（プロフィール画面）
      description: |
        プロフィール画面（テーマタグ・配信設定・統計）の初期表示に使う。フロントエンドは
        `PUT /api/users/{userId}/tags` や `PUT /api/users/{userId}/settings` を呼ぶ前に、
        まずこれで現在の設定値を取得してフォームに反映する。
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: "`userId` が存在しない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: USER_NOT_FOUND
                  message: "指定されたユーザーが見つかりません"
...
  /api/users/{userId}/programs/latest/regenerate:
    post:
      operationId: regenerateLatestProgram
      tags: [programs]
      summary: 今日の番組をリセットする（プロフィールの設定、新しい番組を1本追加生成する）
      description: |
        保存済みの `tags` / `lengthMinutes` を使って番組をもう1本生成し、履歴に追加する。

        **改訂**：以前は既存の最新番組を新しい生成結果で置き換える仕様だったが、
        「デモ用に番組をもう1本追加生成する」エンドポイント（`POST /api/users/{userId}/programs`、
        `createAdditionalProgram`）と実質的に同じ処理になったため統合し、こちらが正式な
        実装として残った（置き換えず追加する）。既存の最新番組は削除されない。
        `GET /api/users/{userId}/programs/latest` は作成日時が最も新しい番組を返すため、
        結果としてホーム画面に表示される番組はこの新しい番組に変わる。

        `POST /api/demo/generate` とは異なり、こちらは**永続化する**。パイプライン全体
        （記事取得→LLM→TTS）を再実行するため負荷が高く、**1ユーザーにつき1日3回まで**に
        制限する。回数は `users.reset_count` で管理し、`users.reset_date`（配信時刻を
        跨いだ日付）が変わった時点でサーバー側が自動的に0へリセットする
        （[06-storage.md](../backend/docs/pipeline/06-storage.md)参照）。

        **改訂：非同期実行になった。** 以前はクライアントの接続を保持したまま同期的に
        パイプライン全体を実行していたが、これだとページのリロードやタブを閉じる操作で
        クライアント側のリクエストが中断されると、サーバー側の`r.Context()`もキャンセル
        されて生成処理自体が実際に打ち切られてしまっていた（実運用で確認済みの不具合）。
        現在は、`userId`の存在確認と多重リクエストチェックのみ同期的に行った上で`202`を
        即座に返し、実際のパイプライン実行はリクエストと切り離したバックグラウンドで行う。
        クライアントは`GET /api/users/{userId}/generation-status`をポーリングして進行状況・
        完了・失敗（失敗時はコード付き）を知る。ポーリング中にページを閉じても、
        バックグラウンドの生成処理自体は継続する。

        **多重リクエストはリクエスト受信時点で即座に拒否する。** 同じ`userId`に対して
        生成処理が実行中の間に2回目のリクエストが来た場合、そのリクエストはパイプラインを
        起動せず`409`（`ALREADY_GENERATING`）を即返す（実装方針の詳細は
        [backend/docs/api-handlers.md](../backend/docs/api-handlers.md#regenerateの多重リクエスト防止)参照）。
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "202":
          description: |
            受理された。生成はバックグラウンドで開始され、この時点ではまだ完了していない。
            進行状況・完了・失敗は`GET /api/users/{userId}/generation-status`をポーリング
            して確認する（失敗時は`generation-status`のレスポンスに`lastError`として、
            以前このエンドポイントが直接返していたのと同じ`code`／`message`が入る：
            `UPSTREAM_FETCH_FAILED`・`UPSTREAM_LLM_FAILED`・`UPSTREAM_TTS_FAILED`・
            `ARTICLE_CACHE_EMPTY`・`RESET_LIMIT_EXCEEDED`・`INTERNAL_ERROR`）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegenerateAccepted"
        "404":
          description: "`userId` が存在しない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: USER_NOT_FOUND
                  message: "指定されたユーザーが見つかりません"
        "409":
          description: 同じユーザーの作り直しリクエストがすでに処理中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: ALREADY_GENERATING
                  message: "前回のリクエストを処理中です。しばらく待ってから再度お試しください"
...
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    RegenerateAccepted:
      type: object
      required: [acceptedAt]
      properties:
        acceptedAt:
          type: string
          format: date-time
          description: "生成をバックグラウンドで開始した時刻。完了はまだしていない（GET .../generation-statusで確認する）"
          example: "2026-09-07T21:32:17+09:00"
...
    UserProfile:
      type: object
      required: [userId, tags, deliveryTime, lengthMinutes, resetsUsedToday, resetsLimitPerDay, stats]
      properties:
        userId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
          example: ["AI", "京都"]
        deliveryTime:
          type: string
          description: "`HH:MM`形式（24時間表記）"
          example: "06:00"
        lengthMinutes:
          type: integer
          enum: [5, 10, 15]
          example: 10
        resetsUsedToday:
          type: integer
          description: "本日`POST /api/users/{userId}/programs/latest/regenerate`を呼んだ回数"
          example: 0
        resetsLimitPerDay:
          type: integer
          description: 1日あたりの作り直し上限（固定値・現状3）
          example: 3
        stats:
          type: object
          required: [totalPrograms, totalDurationSec]
          properties:
            totalPrograms:
              type: integer
              description: これまでに生成された番組数（全期間）
              example: 12
            totalDurationSec:
              type: integer
              description: |
                生成された番組の合計収録時間（秒）。実際に聴取した時間の計測（再生イベントの
                トラッキング）は行っていないため「合計再生分数」ではなく収録時間の合計である点に注意
              example: 8880
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: "`SCREAMING_SNAKE_CASE`の固定文字列"
              enum:
                - USER_NOT_FOUND
                - INVALID_TAGS
                - INVALID_SETTINGS
                - PROGRAM_NOT_FOUND
                - AUDIO_NOT_FOUND
                - ALREADY_GENERATING
                - RESET_LIMIT_EXCEEDED
                - ARTICLE_CACHE_EMPTY
                - NO_UNSEEN_ARTICLES
                - UPSTREAM_LLM_FAILED
                - UPSTREAM_TTS_FAILED
                - INTERNAL_ERROR
            message:
              type: string
...